+++ R/fetchNASIS_components.R
@@ -2,33 +2,32 @@
 # https://github.com/ncss-tech/soilDB/issues/50
 ## TODO: this will not ID horizons with no depths
 ## TODO: better error checking / reporting is needed: coiid, dmu id, component name
-.fetchNASIS_components <- function(SS=TRUE, rmHzErrors=TRUE, fill = FALSE, stringsAsFactors = default.stringsAsFactors()) {
-  # must have RODBC installed
-  if(!requireNamespace('RODBC'))
-    stop('please install the `RODBC` package', call.=FALSE)
+.fetchNASIS_components <- function(SS = TRUE,
+                                   rmHzErrors = TRUE,
+                                   fill = FALSE,
+                                   stringsAsFactors = default.stringsAsFactors(),
+                                   dsn = dsn) {
+
 
   # ensure that any old hz errors are cleared
   if(exists('component.hz.problems', envir=soilDB.env))
     assign('component.hz.problems', value=character(0), envir=soilDB.env)
-  
+
   # load data in pieces
-  f.comp       <- get_component_data_from_NASIS_db(SS=SS, stringsAsFactors = stringsAsFactors)
-  f.chorizon   <- get_component_horizon_data_from_NASIS_db(SS=SS, fill=fill)
-  f.copm       <- get_component_copm_data_from_NASIS_db(SS=SS, stringsAsFactors = stringsAsFactors)
-  f.cogeomorph <- get_component_cogeomorph_data_from_NASIS_db(SS=SS)
-  f.otherveg   <- get_component_otherveg_data_from_NASIS_db(SS=SS)
-  f.ecosite    <- get_component_esd_data_from_NASIS_db(SS=SS, stringsAsFactors = stringsAsFactors)
-  f.diaghz     <- get_component_diaghz_from_NASIS_db(SS=SS)
-  f.restrict   <- get_component_restrictions_from_NASIS_db(SS=SS)
-  
+  f.comp       <- get_component_data_from_NASIS_db(SS = SS, stringsAsFactors = stringsAsFactors, dsn = dsn)
+  f.chorizon   <- get_component_horizon_data_from_NASIS_db(SS = SS, fill = fill, dsn = dsn)
+  f.copm       <- get_component_copm_data_from_NASIS_db(SS = SS, stringsAsFactors = stringsAsFactors, dsn = dsn)
+  f.cogeomorph <- get_component_cogeomorph_data_from_NASIS_db(SS = SS, dsn = dsn)
+  f.otherveg   <- get_component_otherveg_data_from_NASIS_db(SS = SS, dsn = dsn)
+  f.ecosite    <- get_component_esd_data_from_NASIS_db(SS = SS, stringsAsFactors = stringsAsFactors, dsn = dsn)
+  f.diaghz     <- get_component_diaghz_from_NASIS_db(SS = SS, dsn = dsn)
+  f.restrict   <- get_component_restrictions_from_NASIS_db(SS = SS, dsn = dsn)
+
   filled.ids <- character(0)
 
   # optionally test for bad horizonation... flag, and remove
   if(rmHzErrors & nrow(f.chorizon) > 0) {
-    f.chorizon.test <- plyr::ddply(f.chorizon, 'coiid', function(d) {
-      res <- aqp::hzDepthTests(top=d[['hzdept_r']], bottom=d[['hzdepb_r']])
-      return(data.frame(hz_logic_pass=all(!res)))
-    })
+    f.chorizon.test <- aqp::checkHzDepthLogic(f.chorizon, c('hzdept_r', 'hzdepb_r'), idname = 'coiid', fast = TRUE)
 
     # fill=TRUE adds horizons with NA chiid will have NA depths -- will not pass hzDepthTests
     # therefore, only way to use fill effectively was with rmHzErrors=FALSE
@@ -38,36 +37,36 @@
       filled.ids <- as.character(f.chorizon$coiid[filled.idx])
       #print(dput(filled.ids))
     }
-    
+
     # which are the good (valid) ones?
-    good.ids <- as.character(f.chorizon.test$coiid[which(f.chorizon.test$hz_logic_pass)])
-    bad.ids <- as.character(f.chorizon.test$coiid[which(!f.chorizon.test$hz_logic_pass)])
+    good.ids <- as.character(f.chorizon.test$coiid[which(f.chorizon.test$valid)])
+    bad.ids <- as.character(f.chorizon.test$coiid[which(!f.chorizon.test$valid)])
 
     if(length(filled.ids) > 0) {
       good.ids <- unique(c(good.ids, filled.ids))
       bad.ids <- unique(bad.ids[!bad.ids %in% filled.ids])
     }
-    
+
     # keep the good ones
     f.chorizon <- f.chorizon[which(f.chorizon$coiid %in% good.ids), ]
 
     # keep track of those components with horizonation errors
     #if(length(bad.ids) > 0) # AGB removed this line of code b/c it prevents update of 'component.hz.problems' on subsequent error-free calls
     assign('component.hz.problems', value=bad.ids, envir=soilDB.env)
-  } 
-  
+  }
+
   if(nrow(f.chorizon) > 0) {
     # upgrade to SoilProfilecollection
     depths(f.chorizon) <- coiid ~ hzdept_r + hzdepb_r
   } else {
     stop("No horizon data in NASIS component query result.", call.=FALSE)
   }
-  
+
   # add site data to object
   site(f.chorizon) <- f.comp # left-join via coiid
-  
+
   ## TODO: convert all ddply() calls into split() -> lapply() -> do.call('rbind')
-  
+
   # join-in copm strings
   ## 2017-3-13: short-circuts need testing, consider pre-marking mistakes before parsing
   pm <- plyr::ddply(f.copm, 'coiid', .formatcoParentMaterialString, name.sep=' & ')
@@ -94,7 +93,7 @@
 
   # add diagnostic features to SPC
   diagnostic_hz(f.chorizon) <- f.diaghz
-  
+
   # add restrictions to SPC
   # required new setter in aqp SPC object (AGB added 2019/12/23)
   restrictions(f.chorizon) <- f.restrict
@@ -103,7 +102,7 @@
   if(exists('component.hz.problems', envir=soilDB.env))
     if(length(get("component.hz.problems", envir = soilDB.env)) > 0)
       message("-> QC: horizon errors detected, use `get('component.hz.problems', envir=soilDB.env)` for related coiid values")
-  
+
   # set NASIS component specific horizon identifier
   if(!fill & length(filled.ids) == 0) {
     res <- try(hzidname(f.chorizon) <- 'chiid')
@@ -113,20 +112,20 @@
       } else {
         warning("cannot set `chiid` as unique component horizon key -- defaulting to `hzID`")
       }
-    } 
+    }
   } else {
     warning("cannot set `chiid` as unique component horizon key - `NA` introduced by fill=TRUE", call.=F)
   }
-  
+
   # set metadata
   m <- metadata(f.chorizon)
   m$origin <- 'NASIS components'
   metadata(f.chorizon) <- m
-  
+
   # set optional hz designation and texture slots
   hzdesgnname(f.chorizon) <- "hzname"
   hztexclname(f.chorizon) <- "texture"
-  
+
   # done, return SPC
   return(f.chorizon)