+++ tests/testthat/test-fetchKSSL.R
@@ -7,9 +7,16 @@ test_that("fetchKSSL() works", {
   skip_on_cran()
 
   ## sample data
-  x <<- fetchKSSL(series='sierra')
-  x.morph <<- fetchKSSL(series='sierra', returnMorphologicData = TRUE)
-  x.morp.simple.colors <<- fetchKSSL(series='sierra', returnMorphologicData = TRUE, simplifyColors = TRUE)
+  x <<- fetchKSSL(series = 'sierra')
+  
+  x.morph <<- fetchKSSL(series = 'sierra',
+                        returnMorphologicData = TRUE,
+                        progress = FALSE)
+  
+  x.morp.simple.colors <<- fetchKSSL(series = 'sierra',
+                                     returnMorphologicData = TRUE,
+                                     simplifyColors = TRUE,
+                                     progress = FALSE)
 
   # standard request
   expect_true(inherits(x, 'SoilProfileCollection'))
@@ -68,7 +75,7 @@ test_that("fetchKSSL() returns data associated with multiple named series", {
 
   skip_on_cran()
 
-  x.multiple <- fetchKSSL(series=c('sierra', 'amador'))
+  x.multiple <- fetchKSSL(series=c('sierra', 'amador'), progress = FALSE)
   f <- unique(toupper(x.multiple$taxonname)) %in% c('SIERRA', 'AMADOR')
   expect_true(all(f))
 
@@ -97,7 +104,7 @@ test_that("fetchKSSL() fails gracefully when morphology data are missing", {
   # * most lab data
   # * all morphologic data
   # --> cannot simplify colors, so skip
-  res <- suppressMessages(fetchKSSL(pedon_key=37457, returnMorphologicData = TRUE, simplifyColors = TRUE))
+  res <- suppressMessages(fetchKSSL(pedon_key=37457, returnMorphologicData = TRUE, simplifyColors = TRUE, progress = FALSE))
   expect_false(res$morph$phcolor)
   expect_false(res$morph$phfrags)
   expect_false(res$morph$phpores)
@@ -115,11 +122,10 @@ test_that("fetchKSSL() geochem result", {
 
   skip_if_offline()
 
-
   skip_on_cran()
 
   # get geochemical data for a single pedlabsampnum, do some basic filtering
-  res <- fetchKSSL(pedlabsampnum = c("93P0249"), returnGeochemicalData = TRUE)
+  res <- fetchKSSL(pedlabsampnum = c("93P0249"), returnGeochemicalData = TRUE, progress = FALSE)
 
   expect_true(all(filter_geochem(res$geochem, prep_code='S')$prep_code == 'S'))